+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 -------------------------------------------------
 * Added `contEff` and `runContEff`
 * Added `castEff`
+* Added `evalStateEff`
 * Added `Semigroup` and `Monoid` instances for `Match`, `Comp`, `Prod`
 
 0.4.3