+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added `evalStateEff`
 * Added `Semigroup` and `Monoid` instances for `Match`, `Comp`, `Prod`
 * Added `evalStateDef`, `execStateDef`, and `execWriterDef`
+* Added `mkFieldAs`
 
 0.4.3
 -------------------------------------------------